[fix]domain: Month構造体の新規作成メソッドの引数型を修正し、エラーハンドリングを改善

+++ core/src/error.rs
@@ -6,7 +6,7 @@ pub enum DomainError {
     InvalidYear(i32),
 
     #[error("月は1-12の範囲で指定してください: {0}")]
-    InvalidMonth(u32),
+    InvalidMonth(u8),
 }
 
 pub type DomainResult<T> = Result<T, DomainError>;
